Puzzle rework to use APIs
