+++ src/components.d.ts
@@ -20,8 +20,10 @@ declare module 'vue' {
     ChartRadar: typeof import('./components/demo-charts/ChartRadar.vue')['default']
     ChessPuzzle: typeof import('./components/ChessPuzzle.vue')['default']
     Puzzle: typeof import('./components/Puzzle.vue')['default']
-    RankingView: typeof import('./components/RankingView.vue')['default']
+    PuzzleRender: typeof import('./components/PuzzleRender.vue')['default']
     RouterWrapper: typeof import('./components/layout/RouterWrapper.vue')['default']
+    Session: typeof import('./components/Session.vue')['default']
+    SessionCreationForm: typeof import('./components/SessionCreationForm.vue')['default']
     StatsCard: typeof import('./components/StatsCard.vue')['default']
     StopWatch: typeof import('./components/StopWatch.vue')['default']
     TrainingSession: typeof import('./components/TrainingSession.vue')['default']
