+++ src/pages/Training.vue
@@ -1,0 +1,31 @@
+<script setup lang="ts">
+definePage({
+  meta: {
+    icon: 'mdi-numeric-0',
+    title: 'Training Session',
+    drawerIndex: 0,
+  },
+})
+const router = useRouter();
+
+let puzzleColection = JSON.parse(localStorage.getItem('puzzleColection') as string);
+
+if (!puzzleColection) {
+  window.location.href = '/Setup';
+}
+
+</script>
+
+<template>
+  <div class="wrapper full-size">
+    <Session/>
+  </div>
+</template>
+
+<style scoped>
+.full-size {
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
+}
+</style>
