--- src/puzzles/.gitignore
@@ -1,1 +1,0 @@
-ratings_*_original.json
